MFC r360311:
arm64: rockchip: rk805: Use a tailq for the attached regulator

Store the attached regulator in a tailq to later find them in ofw_map.
While here, do not attempt to attach a regulator without a name, a node
might exists but if it doesn't have a name the regulator is unused.


git-svn-id: https://svn.freebsd.org/base/stable/12@362411 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f