MFC r362544:
libdevdctl: Force full match of "timestamp" field name

OpenZFS generates events with a "zio_timestamp" field, which gets mistaken for
"timestamp" by libdevdctl due to imprecise string matching.  Then later it is
assumed a "timestamp" field exists when it doesn't and an exception is thrown.

Add a space to the search string so we match exactly "timestamp" rather than
anything with that as a suffix.

Approved by:    mav (mentor)
Sponsored by:   iXsystems, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@362638 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f