MFC r360345:
bridge: epoch-ification

Run the bridge datapath under epoch, rather than under the
BRIDGE_LOCK().

We still take the BRIDGE_LOCK() whenever we insert or delete items in
the relevant lists, but we use epoch callbacks to free items so that
it's safe to iterate the lists without the BRIDGE_LOCK.

Tests on mercat5/6 shows this increases bridge throughput significantly,
from 3.7Mpps to 18.6Mpps.

MFC after:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/12@362650 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f