MFC r353480: Use event handler in SCTP
Use an event handler to notify the SCTP about IP address changes
instead of calling an SCTP specific function from the IP code.
This is a requirement of supporting SCTP as a kernel loadable module.
This patch was developed by markj@, I tweaked a bit the SCTP related
code.

Submitted by:		markj


git-svn-id: https://svn.freebsd.org/base/stable/12@362848 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f